+++ dataflow/src/main/java/org/checkerframework/dataflow/cfg/playground/ConstantPropagationPlayground.java
@@ -3,6 +3,7 @@ package org.checkerframework.dataflow.cfg.playground;
 import org.checkerframework.dataflow.analysis.ForwardAnalysis;
 import org.checkerframework.dataflow.analysis.ForwardAnalysisImpl;
 import org.checkerframework.dataflow.cfg.visualize.CFGVisualizeLauncher;
+import org.checkerframework.dataflow.cfg.visualize.CFGVisualizeOptions;
 import org.checkerframework.dataflow.constantpropagation.Constant;
 import org.checkerframework.dataflow.constantpropagation.ConstantPropagationStore;
 import org.checkerframework.dataflow.constantpropagation.ConstantPropagationTransfer;
@@ -10,29 +11,25 @@ import org.checkerframework.dataflow.constantpropagation.ConstantPropagationTran
 /** Run constant propagation for a specific file and create a PDF of the CFG. */
 public class ConstantPropagationPlayground {
 
-  /** Do not instantiate. */
+  /** Class cannot be instantiated. */
   private ConstantPropagationPlayground() {
-    throw new Error("do not instantiate");
+    throw new AssertionError("Class ConstantPropagationPlayground cannot be instantiated.");
   }
 
   /**
-   * Run constant propagation for a specific file and create a PDF of the CFG.
+   * Run constant propagation analysis on a file.
    *
-   * @param args command-line arguments, not used
+   * @param args command-line arguments
    */
   public static void main(String[] args) {
 
-    /* Configuration: change as appropriate */
-    String inputFile = "Test.java"; // input file name and path
-    String outputDir = "cfg"; // output directory
-    String method = "test"; // name of the method to analyze
-    String clazz = "Test"; // name of the class to consider
+    // Parse the arguments.
+    CFGVisualizeOptions config = CFGVisualizeOptions.parseArgs(args);
 
     // run the analysis and create a PDF file
     ConstantPropagationTransfer transfer = new ConstantPropagationTransfer();
     ForwardAnalysis<Constant, ConstantPropagationStore, ConstantPropagationTransfer>
         forwardAnalysis = new ForwardAnalysisImpl<>(transfer);
-    CFGVisualizeLauncher.generateDOTofCFG(
-        inputFile, outputDir, method, clazz, true, true, forwardAnalysis);
+    CFGVisualizeLauncher.performAnalysis(config, forwardAnalysis);
   }
 }