+++ dataflow/src/test/java/constantpropagation/ConstantPropagation.java
@@ -1,8 +1,5 @@
 package constantpropagation;
 
-import java.io.FileWriter;
-import java.io.IOException;
-import java.util.Map;
 import org.checkerframework.dataflow.analysis.ForwardAnalysis;
 import org.checkerframework.dataflow.analysis.ForwardAnalysisImpl;
 import org.checkerframework.dataflow.cfg.visualize.CFGVisualizeLauncher;
@@ -21,19 +18,12 @@ public class ConstantPropagation {
 
     String inputFile = "Test.java";
     String method = "test";
-    String clazz = "Test";
+    String clas = "Test";
     String outputFile = "Out.txt";
 
     ConstantPropagationTransfer transfer = new ConstantPropagationTransfer();
     ForwardAnalysis<Constant, ConstantPropagationStore, ConstantPropagationTransfer>
         forwardAnalysis = new ForwardAnalysisImpl<>(transfer);
-    Map<String, Object> res =
-        CFGVisualizeLauncher.generateStringOfCFG(inputFile, method, clazz, true, forwardAnalysis);
-    try (FileWriter out = new FileWriter(outputFile)) {
-      out.write(res.get("stringGraph").toString());
-      out.write("\n");
-    } catch (IOException e) {
-      e.printStackTrace();
-    }
+    CFGVisualizeLauncher.writeStringOfCFG(inputFile, method, clas, outputFile, forwardAnalysis);
   }
 }