+++ framework/tests/all-systems/Issue6019.java
@@ -1,0 +1,6 @@
+// @below-java11-jdk-skip-test
+public class Issue6019 {
+  void test() {
+    var magic = new byte[2];
+  }
+}